Verify tayga interface is up before syncing routes
Summary:
It is possible to run into a race condition when `tayga` is first being initialised where the interface exists but is not yet up. This can cause `fbmeshd` to crash if it tries to sync routes.

Fix by verifying that the `tayga` interface is up before trying to sync routes.

(Note: this ignores all push blocking failures!)

Reviewed By: seanbarker

Differential Revision: D16438365

fbshipit-source-id: 4b47eaa75b644ff59b3c4144a8b73f3db7707935