Remove dependency on openr's stats infrastructure
Summary:
Currently `fbmeshd` depends on `openr`'s stats infrastructure for reporting counters via `ZmqMonitorClient`. This is not ideal as it means the `openr` binary must be running even though it is not otherwise needed for `fbmeshd` to operate.

Remove this dependency by moving to a pull model for stats, which can be accessed via the `meshquery dump_stats` call.

Reviewed By: mariof

Differential Revision: D16418079

fbshipit-source-id: 8098daa3942c1f8c6546bdd8b8234c8c7d8e640f