Remove hardcoded instances of mesh interface
Summary:
Over time, many instances of the literal `"mesh0"` made their way into `fbmeshd`. Although this is the default mesh interface name and is generally unlikely to be changed by a user, it is bad practice to have it be so tightly ingrained in the code.

Abstract it out to the pre-existing `--mesh_ifname` / `FLAGS_mesh_ifname` flag.

Reviewed By: jcard0na

Differential Revision: D16494133

fbshipit-source-id: 5c85d60f4707d2190f282f2eebb24367d68fc635