Add stream API for replicating KvStore
Summary:
Adding stream API to subscribe as well as dump content of KvStore. This is
useful for applications like incremental snoop which needs full dump as well for
identifying incremental changes.

We subscribe to KvStore first before requesting full-dump which will guarantee
that no update will be missed. There is possibility that full dump might already
have an update that is also been delivered through stream, duplicated updates.
The update will be ignored with merge function.

Reviewed By: cenzhao

Differential Revision: D16454827

fbshipit-source-id: cecd49a7a2765a81f4a3237e03ecd9c254a5a321