always try deleting associated sta first
Summary:
When we delete a station to make way for a new one, we need
to be sure to remove the one that has most likely already been
uploaded into the kernel, which would be any that have
associated already.  Ideally, we would be passed the proper
struct cand directly rather than have to search again,
but let's go with this work-around for now to avoid having to
break authsae API once again.

Reviewed By: mariof

Differential Revision: D16809129

fbshipit-source-id: e4103a3ff8b35990b8023ee79e30a3eddb245f14