Fix CMakeLists error breaking marconi builds
Summary:
In this diff: D17215578, the file was renamed but was never changed in `CMakeLists.txt`. This is now causing marconi builds to fail (including in that diff, though they seemed to be ignored and it was landed anyway).

Example here: D17292652 of a diff currently rebased onto master that was working before, but after rebasing fails.

Reviewed By: al8

Differential Revision: D17363900

fbshipit-source-id: c3ae2775f415cacd5c26847d9bac79f3e769b69e