getdeps: fix openssl dep for libevent on macos
Summary:
while testing https://github.com/facebook/watchman/pull/743 on macOS
I noticed that the libevent build failed to find openssl.

openssl is special on macos because apple do not ship the headers.
We already build and depend on openssl for the folly manifest on
macos, so this is really just adding a missing edge.

Reviewed By: simpkins

Differential Revision: D17385053

fbshipit-source-id: 1b688537fef422d81a959fc5749c871b9e868baa