Remove the Watchdog
Summary: Since we're removing ourselves from `openr` and moving to a single event loop, we don't need the watchdog. It'd be better in theory to leave this until we have only one event loop, but we're prioritizing getting out of `openr` as soon as possible since we're blocking them on some things, so this needs to go.

Reviewed By: mariof, ilyacodesFB

Differential Revision: D17402707

fbshipit-source-id: 443e629d10deda2ce50f1911bc57c6f208658c88