Part11: Remove magic number in SparkWrapper for UT for waitEvent
Summary:
As reliability week efforts, this diff removed the magic number in `SparkWrapper`, which is used for spark UT.

Change includes the usage of global timeout value. Since we can receive multiple events, so the timeout of one single event keep unchanged. This global timeout value is the corresponding one to replace

```
for (int = 0; i < 3; ++i)
```

Reviewed By: saifhhasan

Differential Revision: D17718997

fbshipit-source-id: 726b359335d3b84dec2ea621155656f50ac333cb