Fix Spark2 crash in emulation with CHECK
Summary:
Fix Spark2 crash due to `processHandshakeMsg`
We are intermittently observe that emulation tests failed with Spark2 in openr_flap_tests.

The root cause of the crash is inside `processHandshakeMsg()`. We assert that this neighbor is already in tracked state.

When all of the Open/R nodes are restarted, potentially, due to out-of-order packet received, we may observe that handshakeMsg will come at the time when node is NOT aware of this neighbor.

Just ignore the msg. :P

Reviewed By: smaheshm

Differential Revision: D17962729

fbshipit-source-id: 93c39b0aee853dad8516b00f486728234fe54606