fix crashing due to no spf result for neighbor node
Summary:
when we are trace route the previous check ```if (!link->isUp() or linksToIgnore.count(link))  ```is not good enough.
Example:

1-----2
 |       |
 |       |
3-----4

let's suppose link 1 and 2 is overloaded, and node 3 is overloaded. And we get the spfResult from node 1 to all other nodes. When we are trying to trace paths from node 1 to node 3, we will consider node 2 because the link between 3 and 2 is up. So, we will call spfResult[2]. Hence, will cause crash. Because node 2 is not reachable from node 1, and there is no spfResult for node 2 if src node is 1.

This diff, added extra check to avoid such situation.

Reviewed By: saifhhasan

Differential Revision: D17975174

fbshipit-source-id: 8ace0a5c8e936e73d2989be201f6a87b7a3477a6