remove node name prefix
Summary:
This is the follow up diff for T58496006
The main goal is to hide node name prefix "e00.0001." from topology data and application logic, which means user will get  raw node name as is, both from topology data and inside container node as hostname.
We will still need this prefix in emulation infra as part of container name in order to keep multi-env work.

The major change includes:
1. introduce container_name as part of Node attribute, and change all the call site.
2. convert Host, Node, Link to dataclass.
3. fix a bug found in openr which triggered by the node name convention change.

Reviewed By: saifhhasan, tonyqiu1019

Differential Revision: D18928936

fbshipit-source-id: 60799c987a2890c41b5e47e53bcd998c794984c7