Do not advertise expiring keys in finalizeFullSync
Summary:
KvStore identifies to be expire keys as keys with ttl less than `ttlDecr`. When keys
are sent to other KvStore, the ttl is decremented. If remaining ttl is less than
`ttlDecr` then key is not sent.

We enforce this behavior for flooding publication, get & dump APIs, however we
don't do when we finalize full-sync.

This diff also improve `TtlDecrementValue` test to remove sleep calls & make it more deterministic.

Reviewed By: jstrizich

Differential Revision: D19861996

fbshipit-source-id: 07a3cb224b04d8d4430fe32c9029c0cfa419a067