Auto initialize the netlink socket
Summary:
This diff aims at improving usability of netlink socket. User shouldn't
explicitly know how and when to initialize the netlink socket. This is internal
detail of `NetlinkProtocolSocket` and should be done automatically.

With previous change of auto assigning pid by kernel, we can safely initialize
the socket in constructor. Earlier socket must be initialized in their
respective threads to use unique-ID.

Reviewed By: xiangxu1121

Differential Revision: D20950387

fbshipit-source-id: 9a44e8c1b85280eda95d6869ae9d063492ceb5ca