APIs to query routes based on filters
Summary:
Adding more APIs to query routes from kernel based on filters. As of now
filtering of routes is supported on `table`, `protocol`, `address-family` and
`type`.

As of today's implementation kernel only support filtering on `address-family`,
other filters are supported on the receiving side.

NOTE: This exposed bug in UT, in which `buildV4RouteDb` was returning v6 routes
and vice versa.

Reviewed By: jstrizich

Differential Revision: D20956936

fbshipit-source-id: 314b61c7d84c9f8e0756d2aad3de8ced468be113