Bug fix in partial adjacency counting
Summary:
If node loses all of it's adjacencies suddently, then KvStore will be completely
disconnected and will have old adjacency database of other nodes in the network.
This causes the partial adjacencies to be counted from all of the neighbors
towards the current node.

Fix is to count partial adjacencies of the nodes that are reachable.

Reviewed By: steven1327

Differential Revision: D21030355

fbshipit-source-id: 2a9340ba73c9d74ba7954d7d138f9911ed355759