Bug fix in schedule full sync timer
Summary:
Issue - Newly added peers or pending subsequent full sync requests were being
delayed by the full sync timer bug.

Reason - If full sync timer is scheduled we do not send out sync request
immediately, instead we wait for timer to kick in and process the subsequent
request

Bug - Adding two peers, sent out two sync requests immediately. This hits the
limit of max 2 outstanding parallel requests and schedule full-sync timer
with 10s timeout. On receipt of responses almost immediately, timer is set as
it is. Subsequent addition of peer or additional sync request to existing
peer should proceed immediately but they all wait for the timer to kick in.

Fix - On receipt of full-sync response, we either schedule timeout
immediately (if more request to be sent in) or cancel the timeout.

Reviewed By: GirasoleY

Differential Revision: D21057373

fbshipit-source-id: 0a40c16d5f91db2182a64f4a2015d7784447091d