Bug fix for slow Kvstore sync
Summary:
Polling Issue - In following logs, the full sync response is sent immediately from requested node
but processed after some delay. This indicates the in-effective polling to read arrived messages on socket.
{P129052480}

Alas, couldn't reproduce the issue with an example code. However, having this
logic in code to drain all the messages before we hit EAGAIN won't be hurtful,
infact be more useful.

Reviewed By: jstrizich

Differential Revision: D21081229

fbshipit-source-id: ed52e8c242141dfd3328ec5d8099fef12ad388c7