Add interface cache in FibHandler
Summary:
FibHandler is a glue between thrift APIs and NetlinkProtocolSocket. One of the
main job is to convert the thrift representation of route to the netlink
representation and vice versa.

In this diff FibHandler now maintains a cache that is initialized lazily for
interface name to index mapping. It also has a API to get loopback interface
index to create MPLS POP action.

Reviewed By: jstrizich

Differential Revision: D21001656

fbshipit-source-id: 7e2056726d722de50eddd584776802577a675c7b