remove return value of `fromStdOptional(...)`
Summary:
For the second versino of  `fromStdOptional(...)`, it's better to return by value (otherwise we are return reference of local variable).

However, after fbgs, `fromStdOptional(...)` return value is never used. Let's remove it so that it's easier to codemod this function away in the future when we get rid of `apache::thrift::DeprecatedOptionalField(...)`.

Reviewed By: jstrizich

Differential Revision: D21179463

fbshipit-source-id: 301da22da44a7c177eaac5574afd12e1c6410209