Bug fix in KvStore for leak in sent full-sync requests
Summary:
Root-Cause - In the parallel link case, a link flap causes the peer spec update with new command ID (to a new interface). Any inflight request/response for the existing peer over existing socket-ID is cancelled but it is not removed from map that tracks the inflight responses.

There is no actual impact. Because pending sync is for the deleted peers socket-ID. This only happens on turbo devices which have 3 parallel links between devices.

NOTE: On EBB we might have this alert when 04/24 package gets rolled out. If we can pick an image past this diff, it will be resolved. cc xiangxu1121

Reviewed By: jstrizich

Differential Revision: D21267191

fbshipit-source-id: ea23faeed4618698ca1fd5f7e1259c290f28bc82