handle same MPLS label created for different nodes of S199938
Summary:
When a node just came up, it's default mpls label is 101. There is a race condition which could let this node announce it's default label to other neighbors. However, this default label could already be used by another node in the same openr cloud. As a result, it ends up have two mpls routes with same label. After the nodes negotiate the mpls label, it will change to use another unused label. So eventually this duplicate MPLS label will go away. However, this will cause following routes programming issue in FIB module:

A node received two MPLS routes for label 101 in Decision, but Fib will only program one of them because of implementation.
Later, when the duplicate 101 route disappear, because of the logic of findDeltaRoutes(), Decision will send Fib a withdrawal for route 101. Fib withdrawals 101 route.

This issue could be reproduce in openr emulation non-consistantly, because it is an expected race condition and replies on many timers.
Paste of log in emulation to show we see this issue: P129713934

Reviewed By: saifhhasan

Differential Revision: D21264676

fbshipit-source-id: 4cb681f56d1a941731a85b9bb0761ebfa0586b48