add MplsStaticRoutes class to cache information
Summary:
added MplsStaticRoutes. This will be used by BgpRib if it enabled add path and segment routing.

When insert an new entry:
it will inject into it's cache. If this is the first path to a prefix, the bool of return will be true to let BgpRib to announce it to PrefixManager. Otherwise false.

When delete an new entry:
it simply deletes the entry in the cache.

Reviewed By: GirasoleY

Differential Revision: D20848143

fbshipit-source-id: 3bfdabc196b040125cf73cbb1ceddac8f4b2751a