let BgpRib generate static MPLS routes
Summary:
compare to D20541107. The idea is to let BgpRib generate MPLS static routes instead of doing it in PrefixManager.

In high level, the logic is more concentrated in less classes. All BGP related processing/information is done on BgpRib.

The general behavior is like following:
if comes a new path to a new prefix, BgpRib will generate a static MPLS label associated with this prefix. BgpRib will let new path information goes to PrefixManager.

if comes a new path to a existing prefix, BgpRib just attach the new nexthop to existing MPLS routes and publish this route to PrefixManager.

if withdraw a path,  BgpRib will delete corresponding nexthop in MPLS routes for that prefix or delete the whole MPLS route if no more nexthop is there. If we have a backup path to the prefix, we announce it to the PrefixManager. If not, we withdraw this prefix.

Reviewed By: GirasoleY

Differential Revision: D20638886

fbshipit-source-id: 0136ad8f0782679b63e72f66b3b7b76b61b26309