Exhaustive test-cases for NetlinkFibHandler
Summary:
Adding exhaustive tests for NetlinkFibHandler
- Adds trivial tests for static APIs
- Unicast route add/remove/sync
- Mpls route add/remove/sync
- Tests for multiple client

Underneath routes are programmed in FakeNetlinkProtocolSocket but it can as well
work with actual system.

Found some bugs in the code with the help of this UT
- For MplsPop nexthop, we should not set any interface (loopback index is
  internal implementation for netlink based FibHandler. Fboss implements it
  different way).
- Forgot to `return` on invalid protocolId

NOTE: Removed singular version of add/remove MPLS route because there are no
thrift APIs for them. Also cleand up some unused clientId in thrift spec.

Reviewed By: jstrizich

Differential Revision: D21327465

fbshipit-source-id: b99167662552964cd7fb69ce859e6d5f8d5497dc