Clean up build warnings with GCC
Summary:
There were tons of build warning I observed while building with GCC. Took a
chance to clean it up and added TODOs throughout the codebase to address the
same. I am hoping that POCs who touch the code would clean it up eventually.

One interesting case was that - we were looping 16 times to check the same
counters in KvStoreTest.cpp lol

Thrift files were showing the following error and the changes in this diff fixes it
> Uncaptured doctext at on line 6.

Reviewed By: xiangxu1121

Differential Revision: D21486734

fbshipit-source-id: c647f0944a81661a7a071b2f21424ee7ab4fc1e1