Decision change to hanlde self advertised prefix
Summary:
previously if one node is be chosen as one of ECMP nodes for a prefix, Open/R will not program any routes.
This way won't work for if all nodes are announcing the same prefix but we do want to program specific routes for it.
So, in order to let it work, i relaxed the restriction. Now, if the node itself the only one announce it, we will not program routes.
If a prefix is announced by multiple nodes and the current node is one of them. We will still program routes to the prefix if it has prependLabel specified.
1. add nexthops which appeared in corresponding local MPLS static routes.
2. Use results from kspf to program the routes. For these routes, add the corresponding prependLabel in MPLS stack to make sure no routing loops.

Reviewed By: saifhhasan

Differential Revision: D20659140

fbshipit-source-id: 54eb92913a04f97e96c25a0da94e714ece076d5c