Bug fix in KvStore for in-effective polling on peerSyncSock
Summary:
Despite previous fixes for reading messages in loop until EAGAIN. We still see
issue in production where node sends out full-sync requests, response is sent
but it not processed. Later on some link-flap or process-restart event ends up
triggering the callback, which leads to processing of all the bufferred
messages.

P131969114 captures the debugging details
- KvStore TCP sessions are up and alive
- Full-sync request is being sent
- Full-sync response is also sent from neighbor node
- BUT response is not processed

P132219680
- After neighbor restart event, suddenly 310 full-sync responses are processed

The conclusion is in-effective ZMQ polling. The hacky fix is to read messages on
the `peerSyncSock_` every second with a timer. In case if we miss notification
the messages will be ready by scheduled timer.

Thrift migration will address the root-cause as we will get rid of ZMQ
completely from Open/R

Reviewed By: xiangxu1121

Differential Revision: D21799028

fbshipit-source-id: a5421b0f40dcefcc61d8e5a49566b9ff1e0402d1