Fix flaky PrefixAllocator.StaticPrefixUpdate test
Summary:
The busy loops might run out of iteration and terminate pre-maturely before the
prefix allocation change is triggerred.

e.g. Failure that indicates the above reason -
https://www.internalfb.com/intern/testinfra/diagnostics/7599824384514292.562949962113303.1591098947/

Using `folly::Baton` for the synchronization instead of busy loop. This will be
more deterministic and ensure we check only after the update.

Reviewed By: jstrizich

Differential Revision: D21852188

fbshipit-source-id: 68ad4dd0141757e54bb83d5f3bdb6c952a452787