improve getKthPaths Perf
Summary: When tracing paths, once we consider a link once, we don't need to ever look at it again. Either it will be part of an edge-disjoint path or there is no viable edge-disjoint path beyond it. In certain topologies like large grids, this can significantly improve the algorithm's runtime.

Reviewed By: GirasoleY

Differential Revision: D22084708

fbshipit-source-id: dc19454b3716d33a1bfc9c70c4591c710d4200be