Fix OpenNSA 6.5.19 hash as Brcm modified the tarball
Summary:
Koray reported issue with OpenNSA while using build VM and observed that we
don't hit the issue on devserver thanks to downloading the lfs cached file.

The root cause:

Broadcom added License notice and uploaded new 6.5.19 with same name.
This changed the hash though - causing fbcode builder to complain.
Fix it by adjusting the hash.

See the first line in the recursive diff of before vs. after: P134467317

(the broken symlinks is an unrelated existing issue, which we have reported to
Broadcom).

Reviewed By: bkoray

Differential Revision: D22280971

fbshipit-source-id: 781079df426f83901509225156cf77a3966d3301