Retry getting link information in PlatformPublisher
Summary:
PlatformPublisher fatals if it can't get the link information from kernel. It
needs this so that it can translate ifIndex to ifname for subsequent address
updates.

The error happens because of TIMEOUT. Hence applying the retry specifically for
TIMEOUT error only and give up after 3 retries.

Example Error from emulation
{P135505201}

Reviewed By: xiangxu1121

Differential Revision: D22485431

fbshipit-source-id: 668bd4c6c72fdcec34fcfa14ed03800b7d2c5a7d