PrefixAllocator - Populate tag and path_preference for route origination
Summary:
As titled this diff populates new route attributes for originated routes from
`PrefixAllocator` module.

`path_preference` is set to 2000. This is highest value intended to be used in
DC Fabric or EBB network and ensure that Open/R originated route wins over BGP
imported routes. This comply with currently implemented logic in Decision. This
will help normalize the best path selection in Decision across originated as
well as imported routes.

`tag` - This is just for the information

Reviewed By: jstrizich

Differential Revision: D22540556

fbshipit-source-id: b6425724db8ce51a86bb9a2b32c9c74c783725d7