LinkMonitor - Remove static route advertisement feature
Summary:
Static route feature allows the well known routes to be advertised from within
Open/R. These routes are passed as CLI arguments. For all practical purposes we
never used this feature in production.

Removing this feature as we haven't been using it. The new PolicyEngine & Config
can do this in much better way if we need be.

Reviewed By: jstrizich, xiangxu1121

Differential Revision: D22541241

fbshipit-source-id: 1ee3752e12525d9b64f4ce80e7b1892bffee51ad