LinkMonitor - Populate tag and path_preference for route origination
Summary:
As titled this diff populates new route attributes for originated routes from
`LinkMonitor` module.

`path_preference` is set to 2000. This is highest value intended to be used in
DC Fabric or EBB network and ensure that Open/R originated route wins over BGP
imported routes. This comply with currently implemented logic in Decision. This
will help normalize the best path selection in Decision across originated as
well as imported routes.

`tag` - This is just for the information. We set two tags, one to indicate that
it is interface subnet, second one to indicate node/interface information. This
can be useful to observe while debugging

Reviewed By: xiangxu1121

Differential Revision: D22541676

fbshipit-source-id: 7ddf9776a1cf3367be63792209c0eda00ce08216