Increase duration in LongPollFixture.LongPollAdjAdded to incorporate Emulation platform environment
Summary:
We are observing that the UT which mimics long poll API to make sure thrift server returns to client if there is any "adj:" key change is received, fails on emulation platform.
The duration of 50ms is not enough on emulation test platform. It takes close to 350ms. The UT passes on devserver with 50ms threshold.

Reviewed By: xiangxu1121

Differential Revision: D22700959

fbshipit-source-id: f78028d1c8f1db5d0a94ae85d87e380e43af0fd8