Increase duration in LongPollFixture.LongPollAdjAdded to incorporate Emulation platform env
Summary:
We are observing that the UT which mimics long poll API to make sure thrift server returns to client if there is any "adj:" key change is received, fails on emulation platform.
The duration of 350ms is not enough on emulation test platform.  A few runs in the stress test take more than 350ms. So, increasing the duration to 400ms. The UT passes on devserver with 50ms threshold.

Reviewed By: saifhhasan

Differential Revision: D22719687

fbshipit-source-id: 72262b164d3cd30d7407aab8d3f66a13d1d79dac