Unified function retrieve prefix algo and type from the list
Summary:
We had two functions to retrieve type and algorithm. However, we always need to
do run both of them together for computing routes.

This diff defines a function that returns `pair<ForwardingType, ForwardingAlgo>`
and simplifies the implementation to a great extent. This also saves some CPU
cycles (one iteration over the prefixEntries instead of two)

Reviewed By: GirasoleY

Differential Revision: D22882215

fbshipit-source-id: 250f26370c714946cbd0c120b9f28b6055e5f463