Best route selection implementation
Summary:
This diff adds an implementation for best route selection amongs all received
prefix-entries. The selection logic itself is very simple as comparision is on
the numeric values. Especially when compared to existing metrics-vector or BGP
best path selection, the simplicity of representation & implementation shines
out.

Using templates so that we can share the best-path selection logic amongs the
two places
1. Decision for selecting best entries
2. Prefix manager for selecting the best entry for advertising to KvStore

Next diffs
- Integrate best path selection logic in PrefixManager
- Integrate best path selection logic in Decision with knob

Reviewed By: GirasoleY

Differential Revision: D23033110

fbshipit-source-id: bf6d76494acace1952ec166765cf97551e65aa5e