Fixing UT - OpenrCtrlFixture.subscribeAndGetKvStoreFilteredWithKeysTtlUpdateOption
Summary:
After TTL is updated in kv store for an existing key, we sometimes see just the TTL updated entry getting published to a client. Sometimes, existing entries as well as an updated entry getting published. This happens  ~50% of the time during timing issues.

Updating the UT so that we check that the updated entry is published eventually.

Reviewed By: saifhhasan

Differential Revision: D23110326

fbshipit-source-id: f44e6d9a0a8f158258ae3e46364e218e41cb3887