Fix potential race condition in get/push
Summary:
Based on a crash found in production we see crash while settind `data` attribute
on pendingRead when a data is being pushed

```
pendingRead.data = std::forward<ValueTypeT>(val);
```

The crash is segfault and hence it indicates the invalid memory access.
Possibily because the `pendingRead` is no longer a valid reference. The actual
object is hold by `get()` or `getCoro()` API in stack. This means `get()` method
returned and destroyed `PendingRead` object on stack. This can happen when
- `getAnyImpl(...)` enqueues the pending-read and returns
- Meanwhile another thread is in the process of pushing the entry and tries to
  write the data attribute. The boolean flag indicating presence of data
  attribute gets set before actual data is set
- Meanwhile the thread trying to read, will post the baton, learn that data
  attribute is set from boolean flag (though the value is being set) and
  invoke `std::move(pendingRead.data)` which leads to crash in the thread that
  is trying to write data attribute.

The fix
- Instead of relying on the optional field for checking if data is populated by `getAnyImpl()`, return the information explicitly. See the function signature of `getAnyImpl()`

Reviewed By: xiangxu1121

Differential Revision: D23143107

fbshipit-source-id: 95cab812e43189babc0f62ab2de68749a2d4ed48