Program all computed next-hops
Summary:
Background - Open/R was performing shortest path routing. LFA feature was added
to compute additional LFA next-hops. However FIB module only programs the
next-hop with smallest metric. The LFA next-hops are activated when shortest
next-hops are invalidated on interface down event.

However, things got complicated when Open/R started supporting non-shortest path
routing, aka KSPF. With this we added special flag in NextHop object that
indicates FIB to program it even though it is not a smallest metric.

We encountered issue because of FIB programming only subset of computed
next-hops. This was observed when static next-hops with metrics=0 is added.

== Solution / Changes
- Program all computed next-hops. Greatly simplifies lot of things
- LFA next-hops are programmed as well if they're computed. LFA next-hop is
  guaranteed to be loop free as it name mention. This will provide wider ECMP
  group on networks such as EBB and TG.
- Remove unnecessary APIs and code to select next-hops with smallest metric.

Reviewed By: jstrizich

Differential Revision: D23308000

fbshipit-source-id: 16c6e837cb9ade4a89dc45b02f151a9042fcfe92