Decision - Export counter for conflicting prefix entries
Summary:
We allow the routes to be advertised with different forwarding algorithm and
type. However, we expect a unicast route to be advertised with same forwarding
information from all areas in our current context of routing.

Export a counter to indicate the conflicting prefixes. So that we can alert on
it.

Differential Revision: D23600220

fbshipit-source-id: 0c6153a101eb5c98bb4423d852bce80c0d6dab3a