Increasing the threshold for RSS memory usage in SystemMetrics
Summary:
The measured RSS value is shown little higher than the threshold in the unit test.

```Expected: (rssMem1.value() / 1e6) < (100), actual: 108.347 vs 100```

Due to small increase in the value, adjusting the threshold value to 150MB.
Ideally, we need to find out the reason for this increase. However, note that RSS itself can overestimate the actual memory usage of a process if shared libraries are in use.

Reviewed By: yi-xian

Differential Revision: D23804551

fbshipit-source-id: 6b607c510ee8228715e9f245824eb0dc860aa5d0