Stopping Openr Thrift Server after stopping different thrift clients in unit tests
Summary:
We have been seeing Timeout issue during stress test of the unit tests in `OpenrCtrlHandlerTest.cpp`.

Initially we thought that the offending UT was getting stuck. Analysis of logs and stack traces showed that some thread was getting stuck in the thread server.

We also noticed that `OpenrThriftServer` was getting stopped before other OpenR clients. So, thought of closing all the clients first before the thrift server is stopped. It seems that solved the issue and probably all the clients could close gracefully.

Reviewed By: xiangxu1121

Differential Revision: D23948180

fbshipit-source-id: 094bdf190456e0af6f3814100794a66e43077044