Imported Next-Hops should be created with no Open/R Area
Summary:
Currently all next-hops are being created with open/r area as "0". Imported next-hops should have no open/r area.

This is causing issues applying the RibPolicy default_weight to imported next-hops.

The goal of the change is for the following call to createNextHop() to pass a std::nullopt optional area argument.
https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/openr/decision/Decision.cpp?commit=aa7749f25493c1e4518e54da32f2a38e48ac6954&lines=1113

Reviewed By: saifhhasan

Differential Revision: D23990386

fbshipit-source-id: 5bf894dcc4698f2196cca8a4808774f5629c4cf0