Bug Fix - Allow special characters in area name
Summary:
Crash (P144665741) was observed on a cont-canary switch `fsw001.p001.f01.atn6`.
The canary switch accidently started using pod name as area-id because of config
change. This caused the prefix keys to be encoded as
`prefix:fsw001.p001.f01.vll2:vll2.f01.p001:[2401:db00:eef0:b0a::/64]`

The encoding worked fine. But when the prefix is withdrawn the key is decoded to
detect the area-ID. And it crashed while decoding because special characters
such as `.`, `-`, or `_` are not supported in area name decoding.

Reviewed By: jstrizich, xiangxu1121

Differential Revision: D24148292

fbshipit-source-id: 591604c166eb930efe94acac63923760630066cb