Bug Fix - NextHop comparator now treats weight 0 and 1 same
Summary:
`NetlinkFibHandler` when syncing route in Linux first compares it with a route
retrieved from linux. The next-hop weight in route from Linux starts from `0`
and is actual weight is `represented-weight + 1`. The next-hop weight in route
from protocols treats `0` as a special weight (ECMP). And represented weight
reflects the true weight.

This causes the same route from protocol and linux fails the `==` comparator
test. This ends up removing the route and re-installing the new route.

Reviewed By: xiangxu1121

Differential Revision: D24494948

fbshipit-source-id: b72ecc09ed5dfe07b674a0675657e37c64cb4141