Add scaled number of prefixes in openr instances in neighbor discovery script
Summary:
Using APIs to inject 10 prefixes per node (total 6250+ in kvstore). This number will be scaled up even higher (48k+) later.

Updated some APIs which are used for post test validation to cater to needs in this test case.

Main change in this DIFF is that we are injecting prefixes and re-doing the same experiment. Want to observe any changes in the metrics that are emitted.

Reviewed By: tonyqiu1019

Differential Revision: D23982107

fbshipit-source-id: 923a8aafa1d40003d0825cf5a6f39c113b94023b