Tune hold time for initial prefix advertisement
Summary:
== Background
For graceful restart behavior, Open/R implements timer based approach to
re-construct the state and synchronize it in KvStore after restarting. The two
major data types are `Adjacencies` and `Prefixes`.

We had the same timer `2 * Spark-KeepAlive-Time` (4 seconds) for both of these
data type. This used to work well so far in our deployment. However with a
bigger route scale, we can easily shoot this time window for re-constructing the
prefixes.

== Solution
For prefix state we use larger window to re-construct the state.

Reviewed By: GirasoleY

Differential Revision: D24638044

fbshipit-source-id: 037b82300cde4beda57bb77a15523d53f35088e4