PrefixManager - Add support for exception tracing on exception
Summary:
Found a crash in emulation that resulted because of exception in `noexcept`
function. This leads to crash. However as a developer we lose context into the
stack trace that threw the exception.

This diff prints the stack trace that threw the exception along with the
remaining stack from the noexcept function.

Reviewed By: GirasoleY

Differential Revision: D24722400

fbshipit-source-id: 0aeb4a29e0b907bd600a276d2ef95627c9c565f9