remove default area params in clients
Summary:
This diff removes the majority of places where defaultAreaID is used as a default parameter. Instead we force KvStore Clients to specify areaId when interacting with KvStore as most commands are relevant for only one area. The main benefit is to prevent bugs. While writing the diff I uncovered a bug where expired key notification were not populated with the proper area ID to This is a large change but much of it is superficial.

## Major changes:
1. all internal clients must specify an area id for any KvStore API that needs one, no more default parameter
2. The KvStore thrift interface will also require an area to be specified for any call to a nodes which is configured for multi area. For backward compatibility reasons, nodes configured for one area can have the areaID defaulted by the server if they do not specify one

## Cosmetic changes:
1. to make the internal APIs more consistent moved area ID to by first parameter everywhere to allow for other existing default params top remain specified
2. Introduce strong typedef `AreaID` to prevent potential bugs in parameter ordering where multiple strings are expected in a row.

Reviewed By: GirasoleY, xiangxu1121

Differential Revision: D24510775

fbshipit-source-id: 2d4d20b36e8bbda2ef9689e09b50a4be00deac09