add area backward compatibliity logic for all public APIs
Summary:
Our legacy zmq endpoint has backward compatibility logic to allow graceful transition from default area name to a proper name but the thrift endpoints did not. I put this logic into a helper function to be called by the zmq endpoint as well as the thrift methods.

I put server logic into a try block and throw an error if the area specified is not correct. we catch that error and return it to the caller via `semiFuture::setException` and in most cases ultimately via thrift. I think this is a useful structure as going forward there may be other input errors that we want to notify the client about using the same pipeline and not crash our server.

Reviewed By: GirasoleY

Differential Revision: D24689404

fbshipit-source-id: 807e41a950f02d398fdba0e8fd89ce69fca9eeb2