Timestamp publications
Summary:
To measure convergence time of kvStore across the area, we will be using this timestamp to record when a node converged to the new keys. This will be much closer to the timestamp we can mark on kvstore monitor server as that has additional hop and buffering overheads before we can drain and timestamp etc.

(Note: this ignores all push blocking failures!)

Reviewed By: saifhhasan

Differential Revision: D25083490

fbshipit-source-id: ac63e81ea395c9dd7a40acc81e92850cb9cf24bb