remove backward compatibility for full prefix dbs
Summary:
When we began switching to per prefix keys, we added this backward compatibility logic to make sure we could toggle the feature off and on in the network without creating route churn or worse, some unexpected state. This is a bit inefficient and no longer necessary as we only advertise per prefix keys now. Removing this rids us of what is now useless code and also improves performance in the decision pipeline quite a bit as you scale up the number of prefixes.

The complexity of this change mostly comes in the unit tests. many were written assuming on `PrefixDatabase` entry in the `KvStore`. here I modify Decision tests that read `KvStore` advertisement to all be per prefix. In a follow up diff (D25158051) I deal with `SpfSolver` tests that modify the `PrefixState` structure directly.

Reviewed By: GirasoleY, xiangxu1121

Differential Revision: D25137112

fbshipit-source-id: 75af0ef168d05ac2b2c339231c03662f7d2e203f