remove deprecated PrefixState::updatePrefixDatabase()
Summary: In the previous diff I retained `PrefixState::updatePrefixDatabase()` in an effort to keep the change size smaller. The only existing callsites for this are in unit tests Here I remove it and instead of modifying all the existing test that assume a single prefix db, I'm adding helper function in for the tests to use to sync the prefix state for a give node.

Reviewed By: xiangxu1121

Differential Revision: D25158051

fbshipit-source-id: dee21ac76ec41d8d680b094e53fa5a26ed482d2e