Add a thrift API to clear Rib Policy in Decision module
Summary:
Provided a thrift API to clear Rib Policy in Decision module.
Added test cases to verify clear RIB policy API functionality.

When the clear API is invoked, `rebuildRoutes()` is triggered which evaluate shortest paths for the routes. We reset `ribPolicy_` in Decision and as a result no rib policy evaluation happens. So, nexthop weights do not get non-zero weights.

Reviewed By: saifhhasan

Differential Revision: D24704128

fbshipit-source-id: 7a14afa1ab8508d4ebc0c16c58d5566cb90eab12