Migrate thrift::IpPrefix usage inside Decision/PrefixState into folly::CIDRNetwork
Summary:
Within Open/R's code base, there is a mixed usage of thrift::IpPrefix and folly::CIDRNetwork. The intuition to define data-structure inside *.thrift file is due to the ser/deser requirement to send msg across the network, for instance, the communication between thrift client and server.

However,  IpPrefix is NOT the case. It is ONLY used as inter module communication of Open/R. Due to the legacy ZMQ socket communication between modules in old days, we have to maintain this and convert it to/from folly::CIDRNetwork, which is time-consuming and error-prone.

Goal is to kill the structure of thrift::IpPrefix inside Open/R and migrate all usage tofolly::CIDRNetwork

The following changes in decision/prefix-state are steps towards that goal.

Reviewed By: xiangxu1121

Differential Revision: D25680204

fbshipit-source-id: 368606ad9ab5eec32b1b02f1a015776a1c6042f9