use safer ways in assigning values to thrift fields or using thrift's value_or
Summary:
There are two ways how we assign values to a Thrift struct fields using accessors.

1. myStruct.name_ref() = "foo";
2. *myStruct.name_ref() = "foo";

where `myStruct` is a Thrift struct object and `name` is a field in the struct. The field can also be `optional`.

We should follow approach 1 and if the field is optional, it will be incorrect otherwise.

The changes here are in spark and kvstore related files. Similar changes will be needed in other files.

References:
- Field access in C++ https://www.internalfb.com/intern/wiki/Thrift/FieldAccess/
 - https://fb.workplace.com/groups/1730279463893632/permalink/2184833148438259/

Another issue is related to usage of `value_or`.
`value_or` always makes a copy of the field or thrift object. If FieldType is a complex object, that can be expensive. Therefore avoid using `value_or`.

Reference
- https://fb.workplace.com/groups/2318225308446702/permalink/2679128139023082/

Reviewed By: allwync

Differential Revision: D27032874

fbshipit-source-id: bdb82528764a1661c245f8721da7aaf81d1f064d