Fix flaky test case in KvStoreTest.cpp
Summary:
Fixes:
- Wait till one key appears in store A after it is set in store B within the same area until timeout;
- Set longer timeout in waitForAllPeersInitialized till all peers are in INITIALIZED state;

Guideline:
https://www.internalfb.com/intern/wiki/Replacing-sleep-calls-with-futures-and-condition-variables-flintlint-rule :
`Unless you're doing an end-to-end test (e.g. it is important for the goal of what you're testing that you hit the actual backend databases/services), you probably don't want to include a sleep call anywhere in the test.`
Here we indeed want to wait keys are proporgated within all stores, then run the validates.  Ideally, this could be in the category of integration test, rather than unit test.

Reviewed By: yi-xian, xiangxu1121

Differential Revision: D27175151

fbshipit-source-id: 91574bc6a42ec089c7fab9848068048188eedf0d