Add backward compatibility for non-default area onboarding in KvStore Thrift Sync
Summary:
When onboarding non-default areaId for switches in DC, openR will restart and try full sync with peer switches still in default areaId. The full sync should succeed to make the onboarding successful. As a result, backward compatibility for non-default areaId onboarding needs to be supported in KvStore Sync.

Today, the backward compatibility exists in ZMQ sync (to be deprecated soon) but not Thrift sync. This diff adds the backward compatibility of full sync for Thrift sync.

Refer to https://fb.quip.com/7gaOAMn85TyW for more details.

Reviewed By: xiangxu1121

Differential Revision: D27391389

fbshipit-source-id: 7f57119020486f3d0d12e34dcae6b19cca007c5b