Add a check in config parsing for non-empty domain name
Summary:
Adding a check in config parsing so that empty domain name is not accepted.
We have a check in Spark constructor for non-empty domain name. We need to prevent OpenR from crashing.

Functionality wise, we use domain name for backward compatibility for now.

In the next diff, we will handle the case where a message arrives from a peer and it has an empty domain name.

Reviewed By: saumitrp

Differential Revision: D27787180

fbshipit-source-id: b4717e316854a372c0b11bf8f75f9cd37acceb46