feat: add support for compiling open/r on 64-bit arm linux
Prior to this patch it was not possible to run open/r on linux running on ARM due to cmake
and openssl being harded to download/compile for the x86_64 arch.

In order to prevent such problem this patch actively compiles cmake
instead of using pre-compiled binaries and also allows the OpenSSLBuilder
to provide the correct build args to openssl,
thus not trying to compile or run x86_64 software.